+++ contracts/assets/BasicAttentionToken.sol
@@ -12,7 +12,7 @@ contract BasicAttentionToken is PreminedAsset {
     // Constant token specific fields
     string public constant name = "Basic Attention Token";
     string public constant symbol = "BAT";
-    uint public constant decimals = 18;
+    uint8 public constant decimals = 18;
     uint public constant preminedAmount = 10**28;
 
     // NON-CONSTANT METHODS