+++ contracts/assets/BitcoinToken.sol
@@ -12,7 +12,7 @@ contract BitcoinToken is PreminedAsset {
     // Constant token specific fields
     string public constant name = "Bitcoin Token";
     string public constant symbol = "BTC-T";
-    uint public constant decimals = 8;
+    uint8 public constant decimals = 8;
     uint public constant preminedAmount = 10**28;
 
     // NON-CONSTANT METHODS