+++ contracts/assets/DogecoinToken.sol
@@ -12,7 +12,7 @@ contract DogecoinToken is PreminedAsset {
     // Constant token specific fields
     string public constant name = "Dogecoin Token";
     string public constant symbol = "DOGE-T";
-    uint public constant decimals = 8;
+    uint8 public constant decimals = 8;
     uint public constant preminedAmount = 10**28;
 
     // NON-CONSTANT METHODS