+++ contracts/assets/GolemToken.sol
@@ -12,7 +12,7 @@ contract GolemToken is PreminedAsset {
     // Constant token specific fields
     string public constant name = "Golem Network Token";
     string public constant symbol = "GNT";
-    uint public constant decimals = 18;
+    uint8 public constant decimals = 18;
     uint public constant preminedAmount = 10**28;
 
     // NON-CONSTANT METHODS