+++ contracts/assets/RepToken.sol
@@ -12,7 +12,7 @@ contract RepToken is PreminedAsset {
     // Constant token specific fields
     string public constant name = "Rep Token";
     string public constant symbol = "REP-T";
-    uint public constant decimals = 18;
+    uint8 public constant decimals = 18;
     uint public constant preminedAmount = 10**28;
 
     // NON-CONSTANT METHODS