+++ contracts/assets/StatusToken.sol
@@ -12,7 +12,7 @@ contract StatusToken is PreminedAsset {
     // Constant token specific fields
     string public constant name = "Status Network Token";
     string public constant symbol = "SNT";
-    uint public constant decimals = 18;
+    uint8 public constant decimals = 18;
     uint public constant preminedAmount = 10**28;
 
     // NON-CONSTANT METHODS