+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Tokens: GNO, GNT, ICN, ANT, BAT, BNT, SNT, ETC, LTC, DOGE, AVT, XRP
+- Second way to subscribe and redeem using referenceAsset directly in Vault
+- CryptoCompare price feed contract
 
 ### Changed
 - Refactor and re-write tests to have better coverage
@@ -14,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Fixed some code to use async/await instead of promises
 - Increase amount of premined token
+- From SafeMath contract to SafeMath library
 
 ## [0.1.3] - 2017-06-13
 ### Added